toolchain.inc: Update and fix the test-toolchain target
Due to some change, the test-toolchain was no longer working, and was
always reporting that the toolchain is out of date.

This fixes the failure, and prints both the expected versions of Clang,
GCC, and IASL on failures.

Additional changes fix some indentation issues and skip trying to update
submodules when the test is run.

Signed-off-by: Martin Roth <martin@coreboot.org>
Change-Id: Ia350f279c3fd3533523996327cc6b2304e0bead4
Reviewed-on: https://review.coreboot.org/c/coreboot/+/48903
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Patrick Georgi <pgeorgi@google.com>